requirements.txt
catboost==1.2.1
fastapi==0.96.0
joblib==1.1.0
loguru==0.7.0
opentelemetry-api==1.19.0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-sdk==1.19.0
pandas==1.4.4
Pillow==9.5.0
prometheus-client==0.17.1
pytest==7.1.2
python-multipart==0.0.6
requests==2.31.0
uvicorn[standard]==0.22.0